Added the load_image function
Hey guys, I added a load_image function to the play.h file so that it
can be used in play.cpp to load the images to the SDL surfaces. If it
needs to be moved, it shouldn't be to difficult, since I commented it
out pretty heavily. I believe the 0, 0, 0 color should serve as a
replacement for the alpha channel.