Reworked the game_init() functions
Fixed a few problems with the Surfaces, and added included the play.h
file into the credits to use the load_image function there.
Added the initialization of SDL, and the simplistic setup of the screen
to the MyGame.cpp file. That will probably need to change, but I figured
I'd sneak it in for now as a place holder.
I also reworked the menu Surfaces to use the load_image function as
well. Also added SDL to the closing sequences.