SPRITE in SDL
Okay, so I'm starting to try and transfer over the sprites to SDL, so
that I can start adding in the rendering calls. This might take me a
while, its looking like it'll be more complicated than I thought. If
either of you would like to take a crack at this, I wouldn't say no to
the help! I was originally going to put it in MyGame, but decided on
play, because it'll keep the file structure from getting quite as messy.