SPRITE is a struct again
The class from the lazy foo tutorial was making this sooo much more
difficult, so I went back to a struct. Kept it in the play file though.
I also moved over the Sprite_Animate function to play.h as well. I don't
think there is anything in it that is DirectX dependent, so I think itll
work for SDL as well.