Missing Headers and bool for bullets
Okay, so I took a look at what Nick has been working on, and I have to
say, a lot of this is a little beyond my experience with SDL (which has
only been this class). But I've done what I can at the moment, and Ill
get in touch with Nick directly so I can figure out what else I can do
to help get this game working.
Added headers in the audio class files that were missing, and I set up
the bool to control when to animate the bullet